+++ src/jobs/mongo/MeetingSchedulingJobMongo.ts
@@ -3,9 +3,10 @@
 // SPDX-License-Identifier: MPL-2.0
 ///////////////////////////////////////////////////////////////////////////////
 import { MeetingSchedulingJob } from "../MeetingSchedulingJob.js";
-import { CalendarEventMongo, MailboxMongo } from "../../mongo.js";
+import { CalendarEventAttendeeLinkMongo, CalendarEventMongo, MailboxMongo } from "../../mongo.js";
 
 export class MeetingSchedulingJobMongo extends MeetingSchedulingJob<CalendarEventMongo> {
     protected calendarEventClass: any = CalendarEventMongo;
     protected mailboxClass: any = MailboxMongo;
+    protected attendeeLinkClass: any = CalendarEventAttendeeLinkMongo;
 }