+++ src/models/mongo/index.ts
@@ -2,6 +2,7 @@ export * from "./AppearancePreferencesMongo.js";
 export * from "./AttachmentMongo.js";
 export * from "./AuditLogEntryMongo.js";
 export * from "./BrandingMongo.js";
+export * from "./CalendarEventAttendeeLinkMongo.js";
 export * from "./CalendarEventMongo.js";
 export * from "./CalendarShareLinkMongo.js";
 export * from "./ContactMongo.js";